+++ relational-db/Makefile
@@ -1,0 +1,8 @@
+build:
+	go build -o bin/main main.go
+
+dev: 
+	./bin/air
+
+run:
+	go run main.go
