+++ relational-db/commons/token.go
@@ -1,0 +1,7 @@
+package commons
+
+import "strings"
+
+func Tokenize(query string) []string {
+	return strings.Split(query, " ")
+}
