--- relational-db/controllers/health.go
@@ -1,10 +1,0 @@
-package controllers
-
-import (
-	"github.com/gin-gonic/gin"
-	"net/http"
-)
-
-func HealthController(c *gin.Context) {
-	c.String(http.StatusOK, "Working!")
-}
