+++ relational-db/health/controller.go
@@ -1,0 +1,11 @@
+package health
+
+import (
+	"net/http"
+
+	"github.com/gin-gonic/gin"
+)
+
+func HealthController(c *gin.Context) {
+	c.String(http.StatusOK, "Working!")
+}
