--- relational-db/parser/token.go
@@ -1,7 +1,0 @@
-package parser
-
-import "strings"
-
-func Tokenize(query string) []string {
-	return strings.Split(query, " ")
-}
